fix file structure and file naming
