+++ documentation/handover/images/base/workshop_authors.png
