+++ src/components/Pages/AllWorkshops/children/WorkshopCard/style.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import * as vars from '../../styles/variables';
+import * as vars from '../../../../../assets/styles/variables'
 
 export const Article = styled.article`
 	margin: 2.5rem 0 0 0;
