+++ src/components/Pages/AllWorkshops/style.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import * as vars from '../../../styles/variables';
+import * as vars from '../../../assets/styles/variables';
 
 export const Main = styled.section`
 	display: flex;
