--- src/components/Pages/pages.stories.js
@@ -1,7 +1,0 @@
-// import React from 'react';
-// import Worksheets from './Worksheets/Worksheets';
-
-export default { title: 'Pages' };
-
-// Needs params as doesn't show on SB
-// export const worksheets = () => <Worksheets />;
