+++ src/components/Pages/workshopOverview/children/OverviewSection.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 
 import * as SC from '../style';
-import PrimaryButton from '../../../buttons/PrimaryButton';
-import SecondaryButton from '../../../buttons/SecondaryButton';
+import PrimaryButton from '../../../shared/Buttons/PrimaryButton'
+import SecondaryButton from '../../../shared/Buttons/SecondaryButton';
 
 export default function Main({ overview, id }) {
 	return (
