+++ src/components/shared/Buttons/PrimaryButton.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useHistory } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import * as vars from '../../styles/variables';
+import * as vars from '../../../assets/styles/variables';
 
 const SC = {
 	PrimaryButton: styled.button`
