+++ src/components/shared/Tag/Tag.stories.js
