+++ src/components/shared/WorkshopHeader/WorkshopHeader.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import TitleBar from './TitleBar';
+import TitleBar from './children/TitleBar';
 import Tag from '../Tag/Tag';
 import * as SC from './style';
