+++ src/components/shared/WorkshopHeader/WorkshopHeader.test.js
