+++ src/components/shared/WorkshopHeader/children/TitleBar.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import * as SC from './style';
+import * as SC from '../style';
 
 export default function TitleBar({ title }) {
 	return (
